fix: heap corruption
This was a tough one, when loading a statically allocated font (from the
header) imgui, assumes it owns the memory and tries to free it when
exiting causing a double free of the mem. Here I fixed it with
dynamically allocating the memory and copying that so imgui can free it.
staticlly allocated memory is freed automatically